+++ util-compose/src/main/kotlin/team/duckie/app/android/util/compose/WindowInsets.kt
@@ -5,7 +5,7 @@
  * Please see full license: https://github.com/duckie-team/ApiLibrary/blob/trunk/LICENSE
  */
 
-package team.duckie.app.android.util.ui
+package team.duckie.app.android.util.compose
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.WindowInsets
